Added immortal enemy, misc.
Added immortal enemy, changed layout of test room, added a couple variables and made some very minor tweaks. The target OS is set to Windows but I think it's a slightly different version than it was set to before, which is why there's over 3000 files listed as changed. It should be as easy to change back as going to File -> Build Settings and replacing the value in the dropdown menu, with no side effects.