Added heart and ammo HUD elements. They update in real time.
Hearts go down when the player takes damage. Ammo is consumed when the player shoots a bullet. The function to refill it when an enemy is hit is present, but the signal isn't sent out in this build so that doesn't appear. But it should appear when merged with the main branch if the signal is sent out there.

Most of the code is in HeartVisual.cs and AmmoVisual.cs, but some edits were made to Player.cs as well to integrate these. Because of this, Player.cs changes will probably need to be added manually instead of merging the files wholesale.